+++ decomp/nmf_methods/serizel.py
@@ -1,5 +1,6 @@
 from ..utils.data import MinibatchEpochIndex
 from ..utils import assertion, normalize
+from .grads import get_gradients
 
 
 _JITTER = 1.0e-15
@@ -13,22 +14,7 @@ def solve(y, D, x, tol, minibatch, maxiter, method,
     Serizel, R., Essid, S., & Richard, G. (2016).
     MINI-BATCH STOCHASTIC APPROACHES FOR ACCELERATED MULTIPLICATIVE UPDATES IN NONNEGATIVE MATRIX FACTORISATION WITH BETA-DIVERGENCE, 13–16.
     """
-    # mini-batch methods
-    if mask is None:
-        gradients_x = {'l2': _grad_x_l2,
-                       'kl': _grad_x_kl,
-                       'poisson': _grad_x_kl}
-        gradients_d = {'l2': _grad_d_l2,
-                       'kl': _grad_d_kl,
-                       'poisson': _grad_d_kl}
-    else:
-        gradients_x = {'l2': _grad_x_l2_mask,
-                       'kl': _grad_x_kl_mask,
-                       'poisson': _grad_x_kl_mask}
-        gradients_d = {'l2': _grad_d_l2_mask,
-                       'kl': _grad_d_kl_mask,
-                       'poisson': _grad_d_kl_mask}
-
+    gradients_x, gradients_d = get_gradients(likelihood, mask)
     grad_x = gradients_x[likelihood] if grad_x is None else grad_x
     grad_d = gradients_d[likelihood] if grad_d is None else grad_d
 
@@ -154,7 +140,7 @@ def solve_asag_mu(y, D, x, tol, minibatch, maxiter,
     return maxiter, D, x
 
 
-def solve_gsag_mu(y, D, x, tol, minibatch, maxiter, 
+def solve_gsag_mu(y, D, x, tol, minibatch, maxiter,
                   mask, rng, xp, grad_x, grad_d, forget_rate):
     """ Algorithm 7 in the paper """
     minibatch_index = MinibatchEpochIndex(len(y), minibatch, rng, xp)
@@ -192,63 +178,3 @@ def solve_gsag_mu(y, D, x, tol, minibatch, maxiter,
         D = D_new
 
     return maxiter, D, x
-
-
-# --- l2 loss ---
-def _grad_x_l2(y, x, d, mask, xp):
-    """ Multiplicative update rule for square loss.
-    Returns Positive (numerator) and negative (denominator).
-    mask is not used.
-    """
-    f = xp.dot(x, d)
-    return xp.dot(y, d.T), xp.dot(f, d.T)
-
-
-def _grad_d_l2(y, x, d, mask, xp):
-    """ update d with l2 loss """
-    f = xp.dot(x, d)
-    return xp.dot(x.T, y), xp.dot(x.T, f)
-
-
-def _grad_x_l2_mask(y, x, d, mask, xp):
-    """ Multiplicative update rule for square loss.
-    Returns Positive (numerator) and negative (denominator).
-    mask is not used.
-    """
-    f = xp.dot(x, d) * mask
-    y = y * mask
-    return xp.dot(y, d.T), xp.dot(f, d.T)
-
-
-def _grad_d_l2_mask(y, x, d, mask, xp):
-    """ update d with l2 loss """
-    f = xp.dot(x, d) * mask
-    y = y * mask
-    return xp.dot(x.T, y), xp.dot(x.T, f)
-
-
-# --- KL loss ---
-def _grad_x_kl(y, x, d, mask, xp):
-    """ Multiplicative update rule for KL loss.    """
-    f = xp.dot(x, d) + _JITTER
-    return xp.dot(y / f, d.T), xp.sum(d.T, axis=0, keepdims=True)
-
-
-def _grad_d_kl(y, x, d, mask, xp):
-    """ update d with KL loss """
-    f = xp.dot(x, d) + _JITTER
-    return xp.dot(x.T, y / f), xp.sum(x.T, axis=1, keepdims=True)
-
-
-def _grad_x_kl_mask(y, x, d, mask, xp):
-    """ Multiplicative update rule for KL loss with mask. """
-    f = xp.dot(x, d) + _JITTER
-    y = y * mask
-    return xp.dot(y / f, d.T), xp.dot(mask, d.T)
-
-
-def _grad_d_kl_mask(y, x, d, mask, xp):
-    """ update d with l2 loss """
-    f = xp.dot(x, d) + _JITTER
-    y = y * mask
-    return xp.dot(x.T, y / f), xp.dot(x.T, mask)
